moved dispatch to Dashboard component, gallery preview now displays artworks from API, organized dashboard folder a little
