+++ frontend/src/components/dashboard/Dashboard.js
@@ -1,5 +1,5 @@
 
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 import clsx from 'clsx';
 import {
   BrowserRouter as Router,
@@ -22,15 +22,18 @@ import {ThemeProvider} from '@material-ui/styles';
 import {Element as ScrollElement} from 'react-scroll';
 import NavigationItems from './NavigationItems';
 import MuseumGallery from './MuseumGallery';
-import GalleryPreview from './GalleryPreview';
+import GalleryPreview from './gallery-components/GalleryPreview';
 import ArtworkCloseUpCard from './ArtworkCloseUpCard';
 import UploadsFields from './UploadsFields';
 import DescLinks from './DescLinks';
-import Gallery from './Gallery';
-import Banner from './Banner';
+import Gallery from './gallery-components/Gallery';
+import Banner from './gallery-components/Banner';
 import LandingPage from './LandingPage';
 import OurMission from './OurMission';
 import ColorImg from './images/colorful.jpeg';
+import {useDispatch, useSelector} from 'react-redux';
+import {fetchMuseumArtworks} from '../../redux/museumArtworkActions';
+import Pagination from '@material-ui/lab/Pagination';
 
 const drawerWidth = 240;
 
@@ -123,6 +126,11 @@ const useStyles = makeStyles(() => ({
   mission: {
     background: theme.palette.secondary.light,
   },
+  pagination: {
+    display: 'flex',
+    justifyContent: 'center',
+    marginTop: theme.spacing(2),
+  },
 }));
 
 export default function Dashboard() {
@@ -136,6 +144,20 @@ export default function Dashboard() {
   const handleDrawerClose = () => {
     setOpen(false);
   };
+
+  // fetches artworks for GalleryPreview and MuseumGallery
+  const artworks = useSelector((state) => (state.museumArtworks.artworks));
+  const dispatch = useDispatch();
+  const limit = 9;
+  const [museumPage, setMuseumPage] = useState(1);
+  const handleChange = (event, value) => {
+    setMuseumPage(value);
+    console.log(value);
+  };
+  useEffect(() => {
+    dispatch(fetchMuseumArtworks(museumPage, limit));
+  }, [dispatch, museumPage]);
+
 
   return (
     <ThemeProvider theme={theme}>
@@ -195,7 +217,11 @@ export default function Dashboard() {
                   <ScrollElement id="OurMission" className={classes.mission}>
                     <OurMission/>
                   </ScrollElement>
-                  <GalleryPreview name="Museum Gallery" link="/museum-gallery"/>
+                  <GalleryPreview
+                    name="Museum Gallery"
+                    link="/museum-gallery"
+                    artworks={artworks.slice(0, 3)}
+                  />
                   <GalleryPreview
                     name="User Uploads Gallery"
                     link="/user-uploads-gallery"
@@ -205,6 +231,14 @@ export default function Dashboard() {
               <Route exact path="/museum-gallery">
                 <Container className={classes.galleryPageWrapper}>
                   <MuseumGallery />
+                  <Container className={classes.pagination}>
+                    <Pagination
+                      count={10}
+                      size="large"
+                      page={museumPage}
+                      onChange={handleChange}
+                    />
+                  </Container>
                 </Container>
               </Route>
               <Route exact path="/upload-artwork">
