+++ frontend/src/components/dashboard/gallery-components/Banner.js
